% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{inspection}
\alias{inspection}
\title{GET /api/v1/inspection/<id> with JSON response}
\usage{
inspection(compendium_id, file = NA, objects = NA, req, res)
}
\arguments{
\item{compendium_id}{required path paramter - the identifier of the compendium}

\item{file}{required query parameter - the relative path to the file which should be opened}

\item{objects}{optional query parameter - names of objects in the binary file to include in the response}

\item{req}{plumber request object}

\item{res}{plumber response object}
}
\description{
GET /api/v1/inspection/<id> with JSON response
}
