% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{See https://github.com/jeroen/jsonlite/issues/62 and https://github.com/jeroen/jsonlite/pull/90}
\usage{
start()
}
\value{
function does not return but starts listening on the defined port and HTTP endpoints
}
\description{
Start the inspecter microservice
}
\details{
All configuration via environment variables.
}
