% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{Start the inspecter microservice}
\usage{
start()
}
\value{
function does not return but starts listening on the defined port and HTTP endpoints
}
\description{
All configuration via environment variables.
}
